persistencia del modo oscuro
