+++ front/src/app/appointments/page.tsx
@@ -1,6 +1,10 @@
 import AppointmentForm from "@/components/AppointmentForm/AppointmentForm";
 import RouteProtect from "@/components/RouteProtect/RouteProtect";
 
+export const metadata = {
+  title: "Agendar Turno | TecniClick",
+};
+
 export default function Page() {
   return (
     <RouteProtect>
