+++ front/src/app/dashboard/page.tsx
@@ -1,6 +1,10 @@
 import DashboardContent from "@/components/DashboardContent/DashboardContent";
 import RouteProtect from "@/components/RouteProtect/RouteProtect";
 
+export const metadata = {
+  title: "Perfil | TecniClick",
+};
+
 export default function DashboardPage() {
   return (
     <RouteProtect>
