+++ front/src/app/providerPremium/page.tsx
@@ -1,6 +1,10 @@
 import Benefits from "@/components/Benefits/Benefits"
 import SubscriptionForm from "@/components/SubscriptionForm/SubscriptionForm"
 
+export const metadata = {
+  title: "Plan Premium | TecniClick",
+};
+
 export default function providerPremium() {
   return (
     <main className="bg-gradient-navbar dark:bg-tertiary max-w-6xl mx-auto py-10 px-4">
