+++ front/src/app/register/page.tsx
@@ -2,6 +2,10 @@ import RegisterProtect from "@/components/RouteProtect/RegisterProtect";
 import RegisterForm from "@/components/RegisterForm/RegisterForm";
 import React from "react";
 
+export const metadata = {
+  title: "Registrarse | TecniClick",
+};
+
 export default function RegisterPage() {
   return (
     <RegisterProtect>
