+++ front/src/components/NavBar/LoginButton.tsx
@@ -34,17 +34,19 @@ const LoginButton = () => {
   return (
     <>
       {isAuthenticated ? (
-        <div className="hidden w-full md:flex justify-between items-center">
+        <div className="hidden md:flex flex-1 justify-evenly items-center">
+          <Link href="/services" className="btn-hundido">Buscar Servicios</Link>
+          <Link href="/contact" className="btn-hundido">Contacto</Link>
           <Link href="/dashboard" className="btn-hundido">Perfil</Link>
-          <button onClick={(event) => logoutHandler(event)} className="px-3 py-1 ml-2 rounded-md bg-secondary hover:bg-quaternary text-primary hover:text-secondary flex justify-center items-center">
-            <span>Cerrar Sesión</span>
-          </button>
+          <button onClick={(event) => logoutHandler(event)} className="px-3 py-1 ml-2 rounded-md bg-secondary hover:bg-quaternary text-primary hover:text-secondary text-center">Cerrar Sesión</button>
         </div>
       ) : (
-        <div className="hidden w-full md:flex justify-end items-center">
-          <Link href="/login" className="text-nowrap hover:text-quaternary">Iniciar Sesión</Link>
-          <div className="px-3 py-[2px] ml-2 rounded-md bg-secondary hover:bg-quaternary text-primary hover:text-secondary flex justify-center items-center">
-            <Link href="/register">Registrarse</Link>
+        <div className="hidden md:flex flex-1 justify-evenly items-center">
+          <Link href="/services" className="btn-hundido">Buscar Servicios</Link>
+          <Link href="/terms" className="hidden xl:flex btn-hundido">Términos y Condiciones</Link>
+          <div className="flex flex-nowrap justify-center items-center">
+            <Link href="/login" className="text-nowrap hover:text-quaternary">Iniciar Sesión</Link>
+            <Link href="/register" className="px-3 py-[2px] ml-2 rounded-md bg-secondary hover:bg-quaternary text-primary hover:text-secondary flex justify-center items-center">Registrarse</Link>
           </div>
         </div>
       )}
